+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Unreleased
+
+* BREAKING: Change source for `GdsApi.worldwide.organisations_for_world_location` method, remove pagination and remove `stub_worldwide_api_has_organisations_for_location` method.
+
 # 89.0.0
 
 * BREAKING: Change source for `GdsApi.worldwide.world_locations` method and remove pagination.